Rename ScriptObjects.* -> ScriptObject.*
