+++ src/scripting/ScriptObject.hpp
