+++ src/scripting/ScriptSymbols.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "ScriptObjects.hpp"
+#include "ScriptObject.hpp"
 #include "ScriptTypes.hpp"
 #include <BsPrerequisites.h>
